move WeChatPayNextAction
