+++ stripe/src/main/java/com/stripe/android/StripePaymentController.kt
@@ -20,6 +20,7 @@ import com.stripe.android.model.Stripe3ds2AuthParams
 import com.stripe.android.model.Stripe3ds2AuthResult
 import com.stripe.android.model.Stripe3ds2Fingerprint
 import com.stripe.android.model.StripeIntent
+import com.stripe.android.model.WeChatPayNextAction
 import com.stripe.android.networking.AlipayRepository
 import com.stripe.android.networking.AnalyticsRequestExecutor
 import com.stripe.android.networking.AnalyticsRequestFactory
